Print Individual Items before sublists
Prior to this commit, the items that were unique to the parent list
would be printed within the rest of the groups that are loaded from the
sublists.

This change ensures that the Individual Items group is always printed at
the topby not adding it to the map formed of the sublists.