Improve Group loading experience
Prior to this commit, the user who was callig load.GroupsRecursively
would have to create a map first and pass it into the function as an
argument. This was a bit strange, forcing the user to be too concerned
with the implementation of the load.

This commit extracts the recursive logic of the function to another
function within the load package, that way, an exposed function just
called load.Groups can be called and it will create the map inside
itself.

For the user, they just call load.Groups and their map will be created
for them.